Refactor render/mesh: remove stubs, extract raycast, fix chunk path
