+++ src/mesh/CMakeLists.txt
@@ -1,7 +1,8 @@
 add_library(mesh STATIC
     mesh.hpp
-    greedy_mesher.hpp
     mesh.cpp
+    greedy_mesher.hpp
+    greedy_mesher.cpp
 )
 
 target_include_directories(mesh PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
