--- src/render/pipeline.cpp
@@ -1,16 +1,0 @@
-#include "pipeline.hpp"
-#include "../core/logging.hpp"
-
-namespace render {
-
-void Pipeline::create() {
-	core::log(core::LogLevel::Info, "Pipeline create (stub)");
-}
-
-void Pipeline::destroy() {
-	core::log(core::LogLevel::Info, "Pipeline destroy (stub)");
-}
-
-} // namespace render
-
-
