--- src/render/pipeline.hpp
@@ -1,13 +1,0 @@
-#pragma once
-
-namespace render {
-
-class Pipeline {
-public:
-	void create();
-	void destroy();
-};
-
-} // namespace render
-
-
